# replace map with slice in Query