```
benchmark                            old ns/op     new ns/op     delta
BenchmarkSimple-4                    6728          6782          +0.80%
BenchmarkBuildTree1000-4             1108448       1131862       +2.11%
BenchmarkBuildTree100000-4           265806935     263615709     -0.82%
BenchmarkQueryTree-4                 80.2          7.33          -90.86%
BenchmarkQuerySerial-4               240040        207910        -13.39%
BenchmarkQueryTreeMax-4              73.0          7.28          -90.03%
BenchmarkQuerySerialMax-4            241232        207559        -13.96%
BenchmarkQueryTreeArray-4            106           103           -2.83%
BenchmarkQuerySerialArray-4          2343672       2104598       -10.20%
BenchmarkEndpoints100000-4           43077896      40116570      -6.87%
BenchmarkInsertNodes100000-4         35941380      35876511      -0.18%
BenchmarkInsertIntervals100000-4     202871438     193046612     -4.84%
```
When most queries return empty or few results, this has a huge benefit
as shown by the benchmarks (10X faster). When queries return many results
we should nearly identical performance as before.